useAuth now has higher level auth concepts
